storagedriver/s3: Fixed a case where Delete was not working when the subpath immediately followed the given path with a "-" and quit early with noop

+++ registry/storage/driver/s3-aws/s3.go
@@ -1063,7 +1063,21 @@ func (d *driver) copy(ctx context.Context, sourcePath string, destPath string) e
 // We must be careful since S3 does not guarantee read after delete consistency
 func (d *driver) Delete(ctx context.Context, path string) error {
 	s3Objects := make([]*s3.ObjectIdentifier, 0, listMax)
-	s3Path := d.s3Path(path)
+
+	// manually add the given path if it's a file
+	stat, err := d.Stat(ctx, path)
+	if err != nil {
+		return err
+	}
+	if stat != nil && !stat.IsDir() {
+		path := d.s3Path(path)
+		s3Objects = append(s3Objects, &s3.ObjectIdentifier{
+			Key: &path,
+		})
+	}
+
+	// list objects under the given path as a subpath (suffix with slash "/")
+	s3Path := d.s3Path(path) + "/"
 	listObjectsInput := &s3.ListObjectsV2Input{
 		Bucket: aws.String(d.Bucket),
 		Prefix: aws.String(s3Path),
@@ -1078,7 +1092,7 @@ func (d *driver) Delete(ctx context.Context, path string) error {
 		// if there were no more results to return after the first call, resp.IsTruncated would have been false
 		// and the loop would exit without recalling ListObjects
 		if err != nil || len(resp.Contents) == 0 {
-			return storagedriver.PathNotFoundError{Path: path}
+			break ListLoop
 		}
 
 		for _, key := range resp.Contents {
@@ -1135,8 +1149,12 @@ func (d *driver) Delete(ctx context.Context, path string) error {
 		}
 	}
 
-	// need to chunk objects into groups of 1000 per s3 restrictions
 	total := len(s3Objects)
+	if total == 0 {
+		return storagedriver.PathNotFoundError{Path: path}
+	}
+
+	// need to chunk objects into groups of 1000 per s3 restrictions
 	for i := 0; i < total; i += 1000 {
 		_, err := s.DeleteObjects(&s3.DeleteObjectsInput{
 			Bucket: aws.String(d.Bucket),
